Speed up input::Context::get_events.
get_events will only return a non-empty result when a device is plugged
in or removed, which means that the vast majority of calls should return
nothing. Avoid the cost of doing a cross-thread IPC by sharing an atomic
count of pending events, so we can do just an atomic read in the common
case.